+++ changelog.md
@@ -1,4 +1,11 @@
-# Complicated Bees 2.3.2
+# Complicated Bees 2.4.0
+
+## Added
+* Ability to output tags as products - #62
+* Dump command - #58
+  * Biome climates
+  * Effect names and descriptions
+  * Default bee stats
 
 ## Fixed
 * Rendering optimizations for bees - #57, #60